bridge interface of VM even if the VM is on a vlan
There are some cases where a VM is on a vlan but it interface
should not be directly attached but it should be bridged
with the local bridge (e.g. the vm container has also
an IP address on that VLAN).

The condition is:
if it is directly attached there should be an interface
named vlan<vlanid> without an IP address otherwise it's
bridged